requirements.txt
Django==5.1.3
djangorestframework==3.15.2
python_dateutil==2.9.0.post0
Requests==2.32.3
django-bootstrap4==24.4
psycopg2-binary==2.9.10
environs==11.1.0